Use earlycon instead of earlyprintk
'earlyprintk' has been removed for the RISC-V architecture since Linux 5.0 [0].
Hence, this option did not have any effect and was simply ignored by the kernel. 'earlycon' is used now instead, however, this still introduces a time frame during booting the linux kernel where kernel messages are not printed to the console when it crashes during this time frame. This is fixed with the 'console' argument.

[0] https://github.com/torvalds/linux/commit/7ba12bb676c2f7848a53224f9a65c6ef256bb3b6

Signed-off-by: Andreas Korb <andreas.korb@aisec.fraunhofer.de>