OpenSBI: Use generic platform instead of fpga/ariane
The generic platform parses the device tree (the fpga/ariane platform does not) and enables the specified ISA extensions. This is helpful for testing new ISA extensions.

The patch also remove the necessity for our OpenSBI patch, which enabled 32-bit builds for the fpga/ariane platform. The generic platform does not require this patch.

Signed-off-by: Andreas Korb <andreas.korb@aisec.fraunhofer.de>